backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy>=2.0.44,<2.1
alembic==1.14.1
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.5
